Align geodata plugin with frontend style guide (#5970)
* refactor(ui): move geodata components into per-component folders
* refactor(ui): split geodata types into per-concern files
* refactor(ui): split geodata constants by concern
* refactor(ui): extract GeodataMode enum into types
* fix(ui): resolve geodata spec TS errors flagged by tscheck
