+++ redisinsight/ui/src/packages/geodata/src/components/GeoHeader/GeoHeader.tsx
