+++ redisinsight/ui/src/packages/geodata/src/components/GeoInspector/GeoInspector.spec.tsx
