+++ redisinsight/ui/src/packages/geodata/src/components/GeoInspector/index.ts
@@ -1,0 +1,1 @@
+export { GeoInspector } from './GeoInspector'
