+++ redisinsight/ui/src/packages/geodata/src/components/GeoSearchVisualization/index.ts
@@ -1,0 +1,1 @@
+export { GeoSearchVisualization } from './GeoSearchVisualization'
