+++ redisinsight/ui/src/packages/geodata/src/components/GeoShapePlot/index.ts
@@ -1,0 +1,1 @@
+export { GeoShapePlot } from './GeoShapePlot'
