+++ redisinsight/ui/src/packages/geodata/src/components/GeoTable/GeoTable.tsx
