+++ redisinsight/ui/src/packages/geodata/src/constants/distance.ts
@@ -1,0 +1,14 @@
+export const DISTANCE_THRESHOLDS = {
+  close: 0.5,
+  middle: 0.85,
+}
+
+export const THRESHOLD_VISIBLE_ZOOM = 12
+
+export const DISTANCE_COLORS = {
+  close: '#008556',
+  middle: '#9c5c2b',
+  far: '#a00a6b',
+}
+
+export const HEAT_COLORS = ['#008556', '#00a382', '#d9822b', '#a00a6b']
