+++ redisinsight/ui/src/packages/geodata/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './geoCommand'
+export * from './plugin'
+export * from './rqeGeo'
